app/core/database.py
# app/core/database.py
import os

from dotenv import load_dotenv
from sqlmodel import SQLModel, Session, create_engine

load_dotenv()

DATABASE_URL = os.getenv("DATABASE_URL")
if not DATABASE_URL:
    raise RuntimeError("DATABASE_URL is not set in .env")

engine = create_engine(DATABASE_URL, echo=True)


def get_session():
    with Session(engine) as session:
        yield session


def create_db_and_tables():
    # ✅ make sure we import app.models, NOT plain "models"
    from app import models  # noqa: F401

    SQLModel.metadata.create_all(engine)
