requirement.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-dotenv
python-jose
argon2-cffi